[release-fix-github-release] Moved release version helpers
