+++ tasks/release.py
@@ -70,10 +70,11 @@ from tasks.libs.releasing.version import (
     VERSION_RE,
     _create_version_from_match,
     current_version,
+    deduce_version,
+    get_version_major,
     next_final_version,
     next_rc_version,
 )
-from tasks.libs.types.version import Version
 from tasks.pipeline import edit_schedule, run
 from tasks.release_metrics.metrics import get_prs_metrics, get_release_lead_time
 
@@ -83,68 +84,6 @@ GITLAB_FILES_TO_UPDATE = [
 ]
 
 BACKPORT_LABEL_COLOR = "5319e7"
-
-
-def deduce_version(ctx, branch, as_str=True, trust=False) -> str | Version:
-    release_version = get_next_version_from_branch(ctx, branch, as_str=as_str)
-
-    print(
-        f'{color_message("Info", Color.BLUE)}: Version {release_version} deduced from branch {branch}', file=sys.stderr
-    )
-
-    if (
-        trust
-        or not os.isatty(sys.stdin.fileno())
-        or yes_no_question(
-            'Is this the version you want to use?',
-            color="orange",
-            default=False,
-        )
-    ):
-        return release_version
-
-    raise Exit(color_message("Aborting.", "red"), code=1)
-
-
-def get_version_major(branch: str) -> int:
-    """Get the major version from a branch name."""
-
-    return 7 if branch == 'main' else int(branch.split('.')[0])
-
-
-def get_all_version_tags(ctx) -> list[str]:
-    """Returns the tags for all the versions of the Agent in git."""
-
-    cmd = "bash -c 'git tag | grep -E \"^[0-9]\\.[0-9]+\\.[0-9]+$\"'"
-
-    return ctx.run(cmd, hide=True).stdout.strip().split('\n')
-
-
-def get_next_version_from_branch(ctx, branch: str, as_str=True) -> str | Version:
-    """Returns the latest version + 1 belonging to a branch.
-
-    Example:
-        get_latest_version_from_branch("7.55.x") -> Version(7, 55, 4) if there are 7.55.0, 7.55.1, 7.55.2, 7.55.3 tags.
-        get_latest_version_from_branch("6.99.x") -> Version(6, 99, 0) if there are no 6.99.* tags.
-    """
-
-    re_branch = re.compile(r"^([0-9]\.[0-9]+\.)x$")
-
-    try:
-        matched = re_branch.match(branch).group(1)
-    except Exception as e:
-        raise Exit(
-            f'{color_message("Error:", "red")}: Branch {branch} is not a release branch (should be X.Y.x)', code=1
-        ) from e
-
-    tags = [tuple(map(int, tag.split('.'))) for tag in get_all_version_tags(ctx) if tag.startswith(matched)]
-    versions = sorted(Version(*tag) for tag in tags)
-
-    minor, major = tuple(map(int, branch.split('.')[:2]))
-
-    latest = versions[-1].next_version(bump_patch=True) if versions else Version(minor, major, 0)
-
-    return str(latest) if as_str else latest
 
 
 @task
